Init dependency verification (#21)
This file is the result of running
  `./gradlew --write-verification-metadata pgp,sha256 --export-keys`
and then adapting the generated metadata file.

We changed the keyformat to armored, so we can have a clear text
database of the keys used to sign dependencies. Because of this we
dropped the binary key file that was generated alongside the *.keys
file.

Next we searched all the keys from the `<ignored-keys>` section via
https://keyserver.ubuntu.com and added all keys we were able to retrieve
from there to the clear text database. A single key remained unknown.
This turned out to be our own signing key which we forgot to upload
to a keyserver. After uploading out key, all key are now part of the
trusted-keys set.

Because keyservers are not meant to be hammered with requests, we turn
of requesting unknown keys from them. This setting however seems to be
removed each time the command from the beginning is run.