Drop more component entries
These two dependencies are available on maven central including signing
signature. They are signed by a key that is already present in the
keyring and trusted-keys section:

  <trusted-key id="AD296CA014321485EB6780FF8B8E0CB0F6A7657E"
group="org.asciidoctor"/>

I verified this by downloading the jar.asc file from maven central, then
extracting the public key information from that file, then downloading
the key from keyserver.ubuntu.com and comparing it to the keys we
already have present in the file. At this point I'm not sure why we had
these two entries left over in the metadata file. Maybe it was just an
oversight, when we started.